+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+### Fixed
+- Automatically limited size of error/help messages for data content validation. #26
+
+
 ## [0.23.0] - 2022-06-18
 ### Added
 - Added Data validation. At the moment it is only implement for Excel formats. #21